Move Linkding UI to linkding Framework Copy configuration UI from main app
