+++ BookmarkCompanion/components/IntegrationDashboardView.swift
@@ -1,0 +1,21 @@
+//
+//  IntegrationDashboardView.swift
+//  BookmarkCompanion
+//
+//  Created by Christian Wilhelm on 06.01.23.
+//
+
+import SwiftUI
+import Linkding
+
+struct IntegrationDashboardView: View {
+    var body: some View {
+        LinkdingDashboardView()
+    }
+}
+
+struct IntegrationDashboardView_Previews: PreviewProvider {
+    static var previews: some View {
+        IntegrationDashboardView()
+    }
+}
