+++ Linkding/ui/dashboard/bookmarks/LinkdingBookmarkTabView.swift
@@ -4,7 +4,6 @@
 //
 
 import SwiftUI
-import Linkding
 import Shared
 
 struct LinkdingBookmarkTabView: View {
