+++ Linkding/ui/dashboard/bookmarks/edit/BookmarkEditor.swift
@@ -1,20 +1,19 @@
 //
-// LinkdingCreateBookmarkView.swift
+// BookmarkEditor.swift
 // Created by Christian Wilhelm
 //
 
 import SwiftUI
-import Linkding
 import Shared
 
-struct LinkdingCreateBookmarkView: View {
+struct BookmarkEditor: View {
     @Environment(\.presentationMode) private var presentationMode
     @EnvironmentObject var bookmarkStore: LinkdingBookmarkStore
     @EnvironmentObject var tagStore: LinkdingTagStore
 
-    @AppStorage(LinkdingSettingKeys.createBookmarkDefaultArchived.rawValue, store: AppStorageSupport.shared.sharedStore) var defaultArchived: Bool = false
-    @AppStorage(LinkdingSettingKeys.createBookmarkDefaultUnread.rawValue, store: AppStorageSupport.shared.sharedStore) var defaultUnread: Bool = false
-    @AppStorage(LinkdingSettingKeys.createBookmarkDefaultShared.rawValue, store: AppStorageSupport.shared.sharedStore) var defaultShared: Bool = false
+    @ObservedObject var bookmark: LinkdingBookmarkEntity
+
+    @AppStorage(LinkdingSettingKeys.syncHadError.rawValue, store: AppStorageSupport.shared.sharedStore) var syncHadError: Bool = false
 
     @State var url: String = ""
     @State var title: String = ""
@@ -70,23 +69,17 @@ struct LinkdingCreateBookmarkView: View {
                     }
                 })
             }
-                .navigationBarTitle("Create Bookmark")
+                .onAppear() {
+                    self.url = self.bookmark.url
+                    self.title = self.bookmark.title
+                    self.description = self.bookmark.urlDescription
+                    self.isArchived = self.bookmark.isArchived
+                    self.unread = self.bookmark.unread
+                    self.shared = self.bookmark.shared
+                    self.tags = Set(self.bookmark.tagNames)
+                }
+                .navigationBarTitle("Edit Bookmark")
                 .toolbar {
-                    ToolbarItemGroup(placement: .navigationBarTrailing) {
-                        Button(action: {
-                            if (self.url != "") {
-                                let repository = LinkdingBookmarkRepository(bookmarkStore: self.bookmarkStore, tagStore: self.tagStore)
-                                let bookmark = repository.createNewBookmark(url: self.url, title: self.title, description: self.description, isArchived: self.isArchived, unread: self.unread, shared: self.shared, tags: self.tags.map{ $0 })
-                                let syncClient = LinkdingSyncClient(tagStore: self.tagStore, bookmarkStore: self.bookmarkStore)
-                                Task {
-                                    try await syncClient.syncSingleBookmark(bookmark: bookmark)
-                                    self.presentationMode.wrappedValue.dismiss()
-                                }
-                            }
-                        }) {
-                            Image(systemName: "tray.and.arrow.down")
-                        }
-                    }
                     ToolbarItemGroup(placement: .navigationBarLeading) {
                         Button(action: {
                             self.presentationMode.wrappedValue.dismiss()
@@ -94,14 +87,24 @@ struct LinkdingCreateBookmarkView: View {
                             Image(systemName: "xmark")
                         }
                     }
+                    ToolbarItemGroup(placement: .navigationBarTrailing) {
+                        Button(action: {
+                            if (self.url != "") {
+                                let repository = LinkdingBookmarkRepository(bookmarkStore: self.bookmarkStore, tagStore: self.tagStore)
+                                let bookmark = repository.updateBookmark(bookmark: self.bookmark, url: self.url, title: self.title, description: self.description, isArchived: self.isArchived, unread: self.unread, shared: self.shared, tags: self.tags.map { $0 })
+                                let syncClient = LinkdingSyncClient(tagStore: self.tagStore, bookmarkStore: self.bookmarkStore)
+                                Task {
+                                    try await syncClient.syncSingleBookmark(bookmark: bookmark)
+                                }
+                                self.presentationMode.wrappedValue.dismiss()
+                            }
+                        }) {
+                            Image(systemName: "tray.and.arrow.down")
+                        }
+                    }
                 }
                 .sheet(isPresented: self.$selectTagsOpen) {
                     SelectTagsView(selectedTags: self.$tags)
-                }
-                .onAppear() {
-                    self.isArchived = self.defaultArchived
-                    self.unread = self.defaultUnread
-                    self.shared = self.defaultShared
                 }
         }
     }
