+++ Linkding/ui/dashboard/bookmarks/filter/LinkdingBookmarkTabFilterView.swift
@@ -4,7 +4,6 @@
 //
 
 import SwiftUI
-import Linkding
 import Shared
 
 struct FilterOption {
