+++ Linkding/ui/dashboard/tags/FilterTagView.swift
@@ -4,7 +4,6 @@
 //
 
 import SwiftUI
-import Linkding
 import Shared
 
 struct FilterTagView: View {
